Speed up page loading times and decrease traffic
To apply this component, add the uk-img attribute to an <img> element and add the data- prefix to the src attribute so that it reads data-src.
Speed up page loading times and decrease traffic by only loading images as they enter the viewport.
<img data-src="" width="" height="" alt="" uk-img>
The width and height attributes are required so an empty placeholder can be generated in the exact same size as the image. Alternatively, use data-width and data-height if you don't want to set the explicit attributes.